+++ lib/dictosaurus.dart
@@ -9,17 +9,15 @@
 ///
 /// Three utility classes provide dictionary and thesaurus and term expansion
 /// functions:
-/// - the [Dictionary] interface exposes methods that return the properties
-///   (definitions, inflections, part-of-speech, phrases, synonyms and
-///   antonyms) of a `term`;
-/// - the [Thesaurus] interface exposes methods that return the synonyms or
-///   antonyms of a `term` ; and
-/// - the [AutoCorrect] interface exposes methods that return alternative
+/// * the [Dictionary] interface exposes methods that return the properties
+///   (etymologies, pronunciations, definitions, inflections, phrases, synonyms
+///   or antonyms) of a `term`, or a translation of a `term`; and
+/// * the [AutoCorrect] interface exposes methods that return alternative
 ///   spellings for a `term` or terms that start with the same characters.
-///
-/// The [DictoSaurus] interface implements the [Dictionary], [Thesaurus] and
-/// [AutoCorrect] interfaces.
-///
+
+/// The [DictoSaurus] interface implements the [Dictionary] and [AutoCorrect]
+/// interfaces.
+
 /// The [DictoSaurus] interface also exposes the [DictoSaurus.expandTerm]
 /// method that performs `term-expansion`, returning a list of terms in
 /// descending order of relevance (best match first). The (expanded) list of
@@ -45,25 +43,21 @@ library dictosaurus;
 
 export 'src/auto_correct/_index.dart'
     show AutoCorrect, AutoCorrectMixin, AutoCorrectBase;
-export 'src/thesaurus/_index.dart'
-    show Thesaurus, ThesaurusMixin, ThesaurusBase;
 export 'src/dictionary/_index.dart'
     show
         Dictionary,
-        DictionaryMixin,
-        DictionaryBase,
         Pronunciation,
         PronunciationMixin,
         PronunciationBase,
-        TermProperties,
-        TermPropertiesMixin,
-        TermPropertiesBase,
+        DictionaryEntry,
+        DictionaryEntryMixin,
+        DictionaryEntryBase,
         TermVariant,
         TermVariantMixin,
         TermVariantBase;
 export 'src/dictosaurus/_index.dart'
     show DictoSaurus, DictoSaurusMixin, DictoSaurusBase;
-export 'src/_common/_index.dart' show DictionaryEndpoint, PartOfSpeech;
+export 'src/_common/_index.dart' show PartOfSpeech;
 export 'package:gmconsult_dart_core/dart_core.dart' show Language;
 export 'package:gmconsult_dart_core/extensions.dart'
     show LanguageStringExtensions;