+++ lib/src/_common/_index.dart
@@ -3,5 +3,4 @@
 // All rights reserved
 
 export 'typedefs.dart';
-export 'dictionary_endpoint.dart';
 export 'part_of_speech.dart';