+++ lib/src/_common/typedefs.dart
@@ -8,12 +8,7 @@ import 'package:gmconsult_dart_core/dart_core.dart';
 
 /// Function definition of an asynchronous callback that returns the properties
 /// of a [term] from a dictionary provider or API.
-///
-/// Optionally specify the [endpoint] to include in the returned
-/// [TermProperties] instance, useful where different API endpoints are
-/// queried for specific properties.
-typedef DictionaryCallback = Future<TermProperties?> Function(String term,
-    [DictionaryEndpoint? endpoint]);
+typedef DictionaryCallback = Future<DictionaryEntry?> Function(String term);
 
 /// Function definition of an asynchronous callback that returns an expansion of
 /// a [term] from an asynchronous provider or API.
@@ -21,6 +16,6 @@ typedef TermExpander = Future<Set<String>> Function(String term,
     [PartOfSpeech? partOfSpeech]);
 
 /// Returns translations for [term] from [sourceLanguage] to as
-/// [TermProperties].
-typedef TranslationCallback = Future<TermProperties?> Function(
+/// [DictionaryEntry].
+typedef TranslationCallback = Future<DictionaryEntry?> Function(
     String term, Language sourceLanguage);