+++ lib/src/_index.dart
@@ -5,7 +5,6 @@
 /// This private barrel file is used internally in code files to expose all the
 /// package objects.  It exports the barrel file from each sub-folder in src.
 export 'auto_correct/_index.dart';
-export 'thesaurus/_index.dart';
 export 'dictionary/_index.dart';
 export 'dictosaurus/_index.dart';
 export '_common/_index.dart';