+++ lib/src/thesaurus/_index.dart
@@ -1,5 +1,0 @@
-// Copyright ©2022, GM Consult (Pty) Ltd.
-// BSD 3-Clause Licensedictosaurus
-// All rights reserved
-
-export 'thesaurus.dart';